resource conflict to develop

+++ backend/Register/Domain/Services/IClientService.cs
@@ -8,6 +8,6 @@ public interface IClientService
     Task<IEnumerable<Client>> ListAsync();
     Task<ClientResponse> CreateAsync(Client client);
     Task<ClientResponse> GetAccount(string email);
-    Task<ClientResponse> GetAccountById(int id);
-    Task<ClientResponse> UpdateAsync(int id, Client client);
+    Task<ClientResponse> GetAccountById(long id);
+    Task<ClientResponse> UpdateAsync(long id, Client client);
 }
